# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "get": {
        "operationId": "login",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "302": {
            "description": "Redirects to Spotify authorization"
          }
        },
        "summary": "Start Spotify OAuth flow",
        "tags": [
          "Api"
        ]
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "callback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "302": {
            "description": "Redirects to frontend after auth"
          }
        },
        "summary": "Handle Spotify OAuth callback",
        "tags": [
          "Api"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get current authenticated user",
        "tags": [
          "Api"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Logout and clear session",
        "tags": [
          "Api"
        ]
      }
    },
    "/auth/token-login": {
      "post": {
        "description": "Use this when you have a Spotify access token but can't use OAuth flow. The token is validated by fetching your Spotify profile.",
        "operationId": "tokenLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in"
          },
          "201": {
            "description": ""
          },
          "401": {
            "description": "Invalid token"
          },
          "403": {
            "description": "Not available in production"
          }
        },
        "summary": "Dev-only: Login with a manually obtained Spotify token",
        "tags": [
          "Api"
        ]
      }
    },
...
    "/admin/messages": {
      "get": {
        "operationId": "getAllMessages",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get all messages (admin only)",
        "tags": [
          "Api"
        ]
      },
      "post": {
        "operationId": "createMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a new message (admin only)",
        "tags": [
          "Api"
        ]
      }
    },
    "/admin/messages/{id}": {
      "patch": {
        "operationId": "updateMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update a message (admin only)",
        "tags": [
          "Api"
        ]
      },
      "delete": {
        "operationId": "deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete a message (admin only)",
        "tags": [
          "Api"
        ]
      }
    }
  },
  "info": {
    "title": "Unpaused Spotify API",
    "description": "The backend for the Spotify guessing game",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "unpaused_session"
      }
    },
    "schemas": {
      "AuthMeResponseDto": {
        "type": "object",
        "properties": {
          "spotifyUserId": {
            "type": "string",
            "example": "spotify_user_123"
          },
          "displayName": {
            "type": "string",
            "example": "John Doe"
          },
          "isTrusted": {
            "type": "boolean",
            "example": false
          },
          "isAdmin": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "spotifyUserId",
          "displayName",
          "isTrusted",
          "isAdmin"
        ]
      },
      "TokenLoginDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Spotify access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "Spotify refresh token (optional)"
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Message Title"
          },
          "note": {
            "type": "string",
            "example": "Message Note"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "note",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Message Title"
          },
          "note": {
            "type": "string",
            "example": "Message Note"
          }
        },
        "required": [
          "title",
          "note"
        ]
      },
      "UpdateMessageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Message Title"
          },
          "note": {
            "type": "string",
            "example": "Message Note"
          }
        }
      }
    }
  }
}